% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_p_value_calculations.R
\name{all_p_value_calculations}
\alias{all_p_value_calculations}
\title{Perform All P-Value Calculations}
\usage{
all_p_value_calculations(object)
}
\arguments{
\item{data}{A list of emmGrid objects from which to extract the contrasts and values.}
}
\value{
A data frame containing the contrasts, positions, and the calculated values.
}
\description{
This function performs all necessary p-value calculations and accompanying calculations
for estimate, SE, df, and t.ratio from a list of emmGrid objects.
}
\examples{
# Example usage:
# data <- list(emmGrid1, emmGrid2, ...)
# result <- all_p_value_calculations(data)
# print(result)
}
