% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_contrasts_result_calculation.R
\name{average_contrasts_result_calculation}
\alias{average_contrasts_result_calculation}
\title{Calculate Average Contrasts Result}
\usage{
average_contrasts_result_calculation(x, data)
}
\arguments{
\item{x}{A character string representing the name of the argument to extract from the emmGrid objects.}

\item{data}{A list of emmGrid objects from which to extract the contrasts and p-values.}
}
\value{
A data frame containing the contrasts, positions, and the average values.
}
\description{
This function calculates the average contrasts result from a list of emmGrid objects.
}
\examples{
# Example usage:
# data <- list(emmGrid1, emmGrid2, ...)
# result <- average_contrasts_result_calculation("p.value", data)
# print(result)

}
