% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_enzyme_ratios.R
\name{calculate_enzyme_ratios}
\alias{calculate_enzyme_ratios}
\title{Calculate Enzyme Activity Ratios}
\usage{
calculate_enzyme_ratios(x, column_name = "median")
}
\arguments{
\item{x}{A list of data frames, where each data frame corresponds to an enzyme's measurements. Each data frame should contain a "sample" column for sample identifiers and a "median" column with median enzyme activity values for each sample. The enzyme data frames should include: Gly, Xyl, Cbh, Pep, Pho, NAG, and L-DOPA.}

\item{column_name}{A character string specifying which column to use for
calculations. The default is "median". Acceptable values are "median" or "mean".}
}
\value{
A data table with the following calculated enzyme activity ratios for each sample:
\itemize{
\item \code{xyl_gly}: Ratio of xylosidase  to glucosidase
\item \code{glu.xyl_cbh}: Sum of glucosidase  and xylosidase , divided by cellobiohydrolase.
\item \code{glu_pep}: Ratio of glucosidase  to leucine aminopeptidas.
\item \code{pep_pho}: Ratio of leucine aminopeptidas to Phosphatase.
\item \code{glu_nag}: Ratio of glucosidase  to N-Acetylglucosaminidase.
\item \code{glu_ldopa}: Ratio of glucosidase to L-DOPA.
\item \code{cbh_ldopa}: Ratio of Cellobiohydrolase to L-DOPA.
\item \code{nag_ldopa}: Ratio of N-Acetylglucosaminidase to L-DOPA.
}
}
\description{
This function computes a variety of enzyme activity ratios from a given data set containing median or mean values of different enzyme activities. It returns a data table with the calculated ratios for each sample, including:
\itemize{
\item Xylanase/Glucuronidase (Xyl/Gly)
\item (Glucuronidase + Xylanase) / Cellobiohydrolase (Gly + Xyl / Cbh)
\item Glucuronidase/Peptidase (Gly/Pep)
\item Peptidase/Phosphatase (Pep/Pho)
\item Glucuronidase/N-Acetylglucosaminidase (Gly/NAG)
\item Glucuronidase/L-DOPA (Gly/L-DOPA)
\item Cellobiohydrolase/L-DOPA (Cbh/L-DOPA)
\item N-Acetylglucosaminidase/L-DOPA (NAG/L-DOPA)
}
}
\examples{
# Assuming 'enzyme_data' is a list of data frames with median enzyme activity values
# enzyme_data <- list(
#   Gly = data.frame(sample = c("S1", "S2"), median = c(1.2, 0.8)),
#   Xyl = data.frame(sample = c("S1", "S2"), median = c(0.5, 0.6)),
#   Cbh = data.frame(sample = c("S1", "S2"), median = c(1.0, 0.9)),
#   Pep = data.frame(sample = c("S1", "S2"), median = c(0.3, 0.4)),
#   Pho = data.frame(sample = c("S1", "S2"), median = c(0.7, 0.8)),
#   NAG = data.frame(sample = c("S1", "S2"), median = c(0.9, 0.5)),
#   Ldopa = data.frame(sample = c("S1", "S2"), median = c(0.4, 0.3))
# )
# result <- calculate_enzyme_ratios(enzyme_data)
# print(result)
}
