% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_date_plotting.R
\name{color_column}
\alias{color_column}
\title{Custom Line Colors for Column Position}
\usage{
color_column()
}
\value{
A scale object for use in ggplot2 plots.
}
\description{
Defines a manual color scale for line colors based on column position, for use in \code{ggplot2} visualizations.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(factor(cyl), color = factor(gear))) +
  geom_line() +
  color_column()
}
