% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzyme_as_data_table.R
\name{enzyme_as_data_table}
\alias{enzyme_as_data_table}
\title{Extracts the enzyme concentration from all the plates' xlsx files for a certain enzyme}
\usage{
enzyme_as_data_table(x, func, ...)
}
\arguments{
\item{func}{enzyme name to be extracted, formtted as read_'enzyme'. Ex; read_glu, read_xyl, read_NAG, read_Pho, read_Cbh, read_L_DOPA and read_Pep.
Check out the function documentations for their explanation and position in the plate.}

\item{...}{}
}
\value{
a df of the selected enzyme extracted from all the plate xlsx files.
}
\description{
Extracts the enzyme concentration from all the plates' xlsx files for a certain enzyme
}
\examples{
 Gly <- enzyme_as_data_table(paths, func = read_glu)
}
