% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_special_values.R
\name{is.inf.data.frame}
\alias{is.inf.data.frame}
\title{Check for Infinite Values in a Data Frame}
\usage{
is.inf.data.frame(x)
}
\arguments{
\item{x}{A data frame to be checked for infinite values.}
}
\value{
A logical data frame where each element is \code{TRUE} if the corresponding element in \code{x} is infinite, and \code{FALSE} otherwise.
}
\description{
This function checks each element of a data frame to determine if it is an infinite value (Inf or -Inf).
It returns a logical data frame of the same dimensions, where \code{TRUE} indicates the presence of infinite values and \code{FALSE} indicates the absence of them.
}
\examples{
df <- data.frame(a = c(1, Inf, -Inf), b = c(3, 2, Inf))
is.inf.data.frame(df)

}
