% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_special_values.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{Check for NaN Values in a Data Frame}
\usage{
\method{is.nan}{data.frame}(x)
}
\arguments{
\item{x}{A data frame to be checked for NaN values.}
}
\value{
A logical data frame where each element is \code{TRUE} if the corresponding element in \code{x} is NaN, and \code{FALSE} otherwise.
}
\description{
This function checks each element of a data frame to determine if it is a NaN (Not a Number).
It returns a logical data frame of the same dimensions, where \code{TRUE} indicates the presence of NaN and \code{FALSE} indicates the absence of NaN.
}
\examples{
df <- data.frame(a = c(1, NaN, 3), b = c(NaN, 2, 3))
is.nan.data.frame(df)

}
