% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_special_values.R
\name{is.zero.data.frame}
\alias{is.zero.data.frame}
\title{Check for Zero Values in a Data Frame}
\usage{
is.zero.data.frame(x)
}
\arguments{
\item{x}{A data frame to be checked for zero values.}
}
\value{
A logical data frame where each element is \code{TRUE} if the corresponding element in \code{x} is zero, and \code{FALSE} otherwise.
}
\description{
This function checks each element of a data frame to determine if it is zero.
It returns a logical data frame of the same dimensions, where \code{TRUE} indicates the presence of zero values and \code{FALSE} indicates the absence of them.
}
\examples{
df <- data.frame(a = c(1, 0, -3), b = c(3, 2, 0))
is.zero.data.frame(df)

}
