% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{n_fun}
\alias{n_fun}
\title{Generate Observation Count for Each Group}
\usage{
n_fun(x)
}
\arguments{
\item{x}{A numeric vector for which the observation count will be computed.}
}
\value{
A data frame with two columns: \code{y} containing the maximum value of \code{x},
and \code{label} containing the number of observations in \code{x}.
}
\description{
This function calculates the number of observations for a given vector \code{x}.
It returns a data frame with the maximum value of \code{x} (used for positioning the label)
and a label that shows the number of observations.
}
\examples{
# Example usage of n_fun
data <- c(1, 2, 3, 4, NA)
n_fun(data)

}
