% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{observation_numbers}
\alias{observation_numbers}
\title{Add Observation Numbers to a ggplot2 Plot}
\usage{
observation_numbers(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mappings created by \code{aes()} in ggplot2.}

\item{data}{The data to be displayed in the plot. If \code{NULL}, the default, the data is inherited from the plot.}

\item{position}{Position adjustment as in \code{geom_text()}.}

\item{na.rm}{A logical value indicating whether missing values (NA) should be removed. Defaults to \code{TRUE}.}

\item{...}{Other arguments passed to \code{stat_summary()}.}
}
\value{
A layer that adds observation numbers to the plot.
}
\description{
This function adds the number of observations for each group as text labels on a ggplot.
It uses the \code{n_fun} function to calculate the observation numbers.
}
\examples{
# Example usage in ggplot2:
library(ggplot2)
ggplot(mtcars, aes(x = factor(cyl), y = mpg)) +
  geom_boxplot() +
  observation_numbers(aes(x = factor(cyl), y = mpg))

}
