% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_enzyme_data.R
\name{read_Cbh}
\alias{read_Cbh}
\title{Read Cellobiohydrolase (Cbh) Enzyme Activity Data}
\usage{
read_Cbh(df)
}
\arguments{
\item{df}{A string representing the path to the Excel file containing the Cbh enzyme activity data.}

\item{range}{anchored(anchor = "Z75", dim = c(2L, 5L), input = NULL, col_names = NULL, byrow = FALSE)}
}
\value{
A data frame containing positive Cbh enzyme activity measurements.
}
\description{
This function reads Cellobiohydrolase (Cbh) enzyme activity measurements from an Excel file.
}
\examples{
# Example usage:
# read_Cbh("enzyme_data.xlsx")
}
