% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_enzyme_data.R
\name{read_glu}
\alias{read_glu}
\title{Read Glucosidase (Glu) Enzyme Activity Data}
\usage{
read_glu(df)
}
\arguments{
\item{df}{A string representing the path to the Excel file containing the Glu enzyme activity data.}

\item{range}{anchored(anchor = "Z51", dim = c(2L, 5L), input = NULL, col_names = NULL, byrow = FALSE)}
}
\value{
A data frame containing positive Glu enzyme activity measurements.
}
\description{
This function reads Glucosidase (Glu) enzyme activity measurements from an Excel file.
}
\examples{
# Example usage:
# read_glu("enzyme_data.xlsx")
}
