% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_L_DOPA.R
\name{read_L_DOPA}
\alias{read_L_DOPA}
\title{Read L-DOPA Measurement Data}
\usage{
read_L_DOPA(df, cells = "Z114:AD114")
}
\arguments{
\item{df}{Path to the Excel file.}

\item{cells}{The cell range in the Excel sheet to read from (e.g., "Z114:AD114").}

\item{col_names}{Names of the columns for the output data frame. Defaults to "rep1", "rep2", ..., "rep5".}
}
\value{
A data frame with positive L-DOPA measurements.
}
\description{
Read L-DOPA Measurement Data
}
\examples{
read_L_DOPA("data.xlsx", cells = "Z114:AD114")
}
