% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_enzyme_data.R
\name{read_Pep}
\alias{read_Pep}
\title{Read Peptidase (Pep) Enzyme Activity Data}
\usage{
read_Pep(df)
}
\arguments{
\item{df}{A string representing the path to the Excel file containing the Pep enzyme activity data.}

\item{range}{anchored(anchor = "Z75", dim = c(2L, 5L), input = NULL, col_names = NULL, byrow = FALSE)}
}
\value{
A data frame containing positive Pep enzyme activity measurements.
}
\description{
This function reads Peptidase (Pep) enzyme activity measurements from an Excel file.
}
\examples{
# Example usage:
# read_Pep("enzyme_data.xlsx")
}
