% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_enzyme_data.R
\name{read_Pho}
\alias{read_Pho}
\title{Read Phosphatase (Pho) Enzyme Activity Data}
\usage{
read_Pho(df)
}
\arguments{
\item{df}{A string representing the path to the Excel file containing the Pho enzyme activity data.}

\item{range}{anchored(anchor = "Z69", dim = c(2L, 5L), input = NULL, col_names = NULL, byrow = FALSE)}
}
\value{
A data frame containing positive Pho enzyme activity measurements.
}
\description{
This function reads Phosphatase (Pho) enzyme activity measurements from an Excel file.
}
\examples{
# Example usage:
# read_Pho("enzyme_data.xlsx")
}
