% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_positive_cells.R
\name{read_positive_cells}
\alias{read_positive_cells}
\title{Read positive cells in the plate
It first checks if the abiotic correction for a cell is positive (saved in the above cell).}
\usage{
read_positive_cells(data)
}
\arguments{
\item{data}{}
}
\value{
a vector of 5 data points correcponding to the 5 concentrations calculated from the plate replicates.
}
\description{
Returns the five replicates that has positive abiotic correction.If an abitoic correction is detected as negative, then it returns 0.
}
