% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_enzyme_data.R
\name{read_xyl}
\alias{read_xyl}
\title{Read Xylosidas (Xyl) Enzyme Activity Data}
\usage{
read_xyl(df)
}
\arguments{
\item{df}{A string representing the path to the Excel file containing the Xyl enzyme activity data.}

\item{range}{anchored(anchor = "Z57", dim = c(2L, 5L), input = NULL, col_names = NULL, byrow = FALSE)}
}
\value{
A data frame containing positive Xyl enzyme activity measurements.
}
\description{
This function reads Xylosidas (Xyl) enzyme activity measurements from an Excel file.
}
\examples{
# Example usage:
# read_xyl("enzyme_data.xlsx")
}
