% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{set_coloring_column}
\alias{set_coloring_column}
\title{This function adds a new column named \code{highlight} to the input data frame. The \code{highlight} column
is generated by combining information from the \code{sample_date} and \code{col_no} columns.}
\usage{
set_coloring_column(data)
}
\arguments{
\item{data}{A data frame that contains at least \code{sample_date} and \code{col_no} columns.}
}
\value{
A data frame with an additional \code{highlight} column. This column is a factor created
by concatenating "before" or "after" (based on \code{sample_date}) with " C" and the fourth character
of \code{col_no}.
}
\description{
This function adds a new column named \code{highlight} to the input data frame. The \code{highlight} column
is generated by combining information from the \code{sample_date} and \code{col_no} columns.
}
\details{
The function checks if the \code{sample_date} column has a value of "0" to assign "before"
and "after" labels. It then appends the fourth character of \code{col_no} to create distinct categories.
}
\examples{
# Example data frame
data <- data.frame(sample_date = c("0", "1", "0", "1"), col_no = c("C123", "C124", "C125", "C126"))
data <- set_coloring_column(data)
}
