% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_pooled_model.R
\name{tidy_pooled_model}
\alias{tidy_pooled_model}
\title{Tidy Pooled Model from MICE Imputation}
\usage{
tidy_pooled_model(lmm)
}
\arguments{
\item{lmm}{A list of linear mixed models (lmm) generated from mice imputation to be tidied and pooled.}
}
\value{
A tibble containing the pooled estimates of the fixed effects from the models.
The tibble includes the following columns:
\itemize{
\item \code{term}: The term or predictor in the model.
\item \code{estimate}: The mean estimate of the term across all models.
\item \code{std.error}: The mean standard error of the term across all models.
\item \code{statistic}: The mean statistic (e.g., t-value) of the term across all models.
\item \code{p.value}: The mean p-value of the term across all models.
}
}
\description{
This function takes a list of linear mixed models generated from mice imputation and returns a tidied and pooled model.
}
\examples{
# Example usage:
# Assuming `imputed_models` is a list of linear mixed models from mice imputation
pooled_results <- tidy_pooled_model(imputed_models)

}
